Conflicts:
	code/rd-vanilla/ghoul2/G2_API.cpp
	code/rd-vanilla/tr_font.cpp
	code/rd-vanilla/tr_ghoul2.cpp
	code/rd-vanilla/tr_init.cpp
	code/rd-vanilla/tr_local.h
	code/rd-vanilla/tr_main.cpp
	code/rd-vanilla/win_glimp.cpp

+++ codemp/client/snd_mem.cpp
@@ -101,7 +101,7 @@ void DumpChunks(void)
 		memcpy (str, data_p, 4);
 		data_p += 4;
 		iff_chunk_len = GetLittleLong();
-		Com_Printf ("0x%x : %s (%d)\n", (int)(data_p - 4), str, iff_chunk_len);
+		Com_Printf ("0x%x : %s (%d)\n", (intptr_t)(data_p - 4), str, iff_chunk_len);
 		data_p += (iff_chunk_len + 1) & ~1;
 	} while (data_p < iff_end);
 }
